pyproject.toml
[project]
name = "mobile-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11.10",
    "aiomysql>=0.2.0",
    "alembic>=1.14.0",
    "fastapi[standard]>=0.115.6",
    "greenlet>=3.1.1",
    "passlib[bcrypt]>=1.7.4",
    "pillow>=11.1.0",
    "pydantic-settings>=2.6.1",
    "pyjwt>=2.10.1",
    "python-decouple>=3.8",
    "pytz>=2024.2",
    "qrcode>=8.0",
    "requests>=2.32.3",
    "sqlalchemy>=2.0.36",
]
